Move various bsp base classes from `solid` module to `bsp`
